requirements.txt
django==2.2.9
djangorestframework==3.11.0
psycopg2-binary==2.8.4
django-pandas==0.6.1
pivottablejs==0.9.0
IPython==7.14.0